+++ feincms3/__init__.py
@@ -1,2 +1,2 @@
-VERSION = (0, 32, 0)
+VERSION = (0, 32, 1)
 __version__ = ".".join(map(str, VERSION))